remove wchar/wstring capabilities
a substantial portion of the code consists in handling windows stupid
wchar.  with this commit, we remove all support for it, de-duplicating
many things.  all the tests pass, everything compiles.

next steps should be: address the TODOs and NOTEs sprinkled throughout
remove all the <CharT> templating as unnecessary.
remove TargetString/TargetStringView (left as a stupid variant)